fix(web): fix RSC dashboard crash + MUI SSR hydration (React #418)
The RSC dashboard passed `component={Link}` to MUI client components,
which throws "Functions cannot be passed directly to Client Components"
during the Server Components render — crashing the page and causing the
React #418 hydration error in production.

- Split the dashboard into a server page (app/page.tsx) that only fetches
  data, and a client island (DashboardView) that owns the interactive
  links/refresh. The server passes the serializable data down; the client
  view freely uses component={Link}.
- Add @mui/material-nextjs AppRouterCacheProvider in the root layout so
  MUI/emotion styles render correctly under SSR/RSC (the canonical fix for
  MUI + App Router hydration mismatches), now that pages actually
  server-render MUI markup.

Verified: prod server returns 200 for / and /version with zero
"Functions cannot be passed" render errors.
