+++ web/package.json
@@ -20,6 +20,7 @@
     "@emotion/styled": "^11.14.1",
     "@mui/icons-material": "^7.3.6",
     "@mui/material": "^7.3.6",
+    "@mui/material-nextjs": "^9.0.1",
     "js-yaml": "^4.1.1",
     "next": "16.2.6",
     "react": "19.2.3",
